add newHouses.js, duplication-localStorage.js style error msg
